chore(phpstan): extend level 9 to the leaf modules (#2403)
* chore(phpstan): extend level 9 to the leaf modules

Generalize the scoped compiler pass into phpstan-strict.neon and add the
small leaf modules (Config, Console, Filesystem, Formatter, HttpClient,
Interop, Lsp, Nrepl, Profile, Watch) to the level-9 set alongside the
compiler. Renames composer phpstan-compiler -> phpstan-strict and the CI
step. Rest of src/ stays at level 6 (phpstan.neon).

Resolves the ~52 findings this surfaces, mostly mechanical:
- Factory facade getters: @var the cross-module facade returned by
  getProvidedDependency() (the established ApiFactory idiom).
- Config value casts: add Phel\Shared\ScalarCoercion (toString/toInt/
  toFloat) to coerce loosely-typed config mixed to a scalar with a
  default, replacing (string)/(int)/(float) casts of mixed across the
  config/command readers.
- Narrow iterator/list/option reads to concrete types (PhelPathFilter
  SplFileInfo, FunctionsToExportFinder FnInterface guard, array_values
  for list returns, is_int/is_array guards).
- ProfilerSession stores an int hrtime; ProfilingFn calls __invoke
  explicitly; AbstractShellWatcher guards a null stdout before fgets.

Two documented ignoreErrors remain (structurally untypeable but safe):
CompilerFacade::encodeNs (Gacela cached() mixed + facadeOnlyDelegates)
and PhelConfig::with() (named-arg spread of a merged array).

No behavior change; unit/integration (4313) and core (5941) suites green.

* style: blank lines after guard clauses (NewlineAfterStatementRector)

+++ src/php/Profile/Domain/ProfilerSession.php
@@ -56,7 +56,7 @@ final class ProfilerSession implements ProfilerHookInterface
      */
     public function enter(string $name): void
     {
-        $this->stack[] = [$name, hrtime(true), 0];
+        $this->stack[] = [$name, (int) hrtime(true), 0];
         ++$this->depth;
     }
